fix : 타임캡슐 조회할 때 몇번째 타임캡슐인지 계산 수정

+++ backend/timecapsule-service/src/main/java/com/familring/timecapsuleservice/dto/response/TimeCapsuleStatusResponse.java
@@ -15,7 +15,7 @@ public class TimeCapsuleStatusResponse {
 
     private int status;
     private int dayCount; // 남은 날짜
-    private Long count; // 몇 번째 타임캡슐 인지
+    private int count; // 몇 번째 타임캡슐 인지
     private List<UserInfoResponse> users; // 작성한 가족 구성원 목록
 
 }
